Version 0.2.2

+++ pharmacy/database/models.py
@@ -14,7 +14,7 @@ class Medicines(models.Model):
     category = models.ForeignKey('Category', on_delete=models.PROTECT)
     name = models.CharField('Название', max_length=255)
     prise = models.FloatField('Цена')
-    count = models.FloatField('Кол-во')
+    count = models.IntegerField('Кол-во')
 
     def __str__(self):
         return self.name
